+++ src/package.json
@@ -16,6 +16,8 @@
     "gulp-sourcemaps": "^1.7.3",
     "moment": "^2.17.1",
     "react": "^15.3.2",
+    "react-dnd": "^2.2.3",
+    "react-dnd-html5-backend": "^2.2.3",
     "react-dom": "^15.3.2"
   },
   "babel": {
@@ -29,8 +31,6 @@
   },
   "devDependencies": {
     "babel-preset-es2015": "^6.18.0",
-    "draft-js": "^0.9.1",
-    "draft-js-export-markdown": "^0.2.2",
     "lodash": "^4.17.2",
     "slate": "^0.16.10"
   }
